fix hydration error by removing root layout
