+++ app/[locale]/[org]/page.tsx
@@ -2,6 +2,7 @@ import { getProjects } from '~/server/queries/projects';
 import CreateProjectForm from '~/app/[locale]/[org]/_components/CreateProjectForm';
 import ProjectCard from './_components/ProjectCard';
 import { routes } from '~/lib/routes';
+import Section from '~/components/layout/Section';
 
 type OrgPageProps = {
   // ✅ Never assume the types of your params before validation
@@ -15,7 +16,7 @@ export default async function OrgPage({ params }: OrgPageProps) {
   return (
     <div className="flex flex-col p-12">
       <h1 className="pb-4 text-4xl">Organization Page</h1>
-      <div className="grid grid-cols-4 space-x-4">
+      <Section>
         {allProjects
           ? allProjects.map((project) => (
               <ProjectCard
@@ -25,7 +26,7 @@ export default async function OrgPage({ params }: OrgPageProps) {
               />
             ))
           : 'No projects found'}
-      </div>
+      </Section>
       <CreateProjectForm orgSlug={org} />
     </div>
   );
