+++ app/[locale]/layout.tsx
@@ -1,5 +1,16 @@
 import { NextIntlClientProvider } from 'next-intl';
+import { Inter } from 'next/font/google';
 import { getMessages } from 'next-intl/server';
+import { Metadata } from 'next';
+import ResponsiveContainer from '~/components/layout/ResponsiveContainer';
+import '../globals.scss';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
 
 export default async function LocaleLayout({
   children,
@@ -14,9 +25,9 @@ export default async function LocaleLayout({
 
   return (
     <html lang={locale}>
-      <body>
+      <body className={inter.className}>
         <NextIntlClientProvider messages={messages}>
-          {children}
+          <ResponsiveContainer>{children}</ResponsiveContainer>
         </NextIntlClientProvider>
       </body>
     </html>
