--- app/layout.tsx
@@ -1,25 +1,0 @@
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-import './globals.scss';
-import ResponsiveContainer from '~/components/layout/ResponsiveContainer';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <ResponsiveContainer>{children}</ResponsiveContainer>
-      </body>
-    </html>
-  );
-}
