--- components/ui/Skeleton.tsx
@@ -1,11 +1,0 @@
-import { cn } from '~/lib/utils';
-
-function Skeleton({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) {
-  const classes = cn('animate-pulse rounded-md bg-platinum', className);
-  return <div className={classes} {...props} />;
-}
-
-export { Skeleton };
